Python/APP1/Python_Bootcamp.Day_06/src/requirements.txt
grpcio
grpcio-tools
googleapis-common-protos
pydantic
sqlalchemy
psycopg2 binary
alembic
postgresql